requirements.txt
spacy
openai
anthropic
tiktoken
tqdm
python-dotenv
datasets
rank_bm25
nltk